requirements.txt
requests>=2.27
mock>=4.0
tox>=2.1.1
responses>=0.20.0
django>=3.0
djangorestframework>=3.1
pytest
urllib3>=1.20
six>=1.10